* Fix lineheight

* Added status tooltip

* custom log prefix if dev

* fix issue with None in when adding repository

* Bump minimum version

* Use LooseVersion

* add template tester

* Add devcontainer

* update gitignore

* move config

* update

* update

* update

* update

* mv

* Changes to dev container

* Use port 8124

* Add devcontainer docs

* fix reload

* Fix cache issue

* Adds custom_component_helper

* update custom_component_helper

* Adds shebang

* Add readme

* restructure tasks

* typo

* typo

* hide log

* Fix versionhandling for beta/dev versions

* update

* Fix hide

* Initialize onboarding

* Correct issue link fixes #324

* Add migration to schema 5 (Issue #322)

* Return 404

* Add new base structure

* Fix issue #327

* Add integrationhelper

* Add new HacsRepository class

* Add initial theme with new structure

* More additions for new format

* more structure updates

* More changes for new structure

* Add plugin to new structure

* Fix plugin issues

* Fix download issue

* It starts!, it's not working, but it starts!

* not completly broken

* Settings mostly work again

* Move AIOGitHub to external lib

* Adds "tabnine.tabnine-vscode" to devcontainer

* Add initial config_flow support

* More config flow updates

* rename

* Fix CF loading when using yaml

* More updates after restructure

* rename

* Fix layout issues

* Reenable CF

* Adds "github.vscode-pull-request-github"

* update

* format

* update

* Adds TabNine Cloud AutoCompleter

* adds pyls

* Format

* update

* Mostly not total shit

* update

* Not too bad

* Fix backup issues

* it kinda works

* Fix restore for hacs

* updates

* Cleanup on isle 3

* typo

* Startup fixes

* Not all the bad

* updates

* updates

* Fail if loading data failed

* download logs

* fix issues

* Fix pending upgrade

* Changes to restore

* create backups

* More validations

* more fixes

* no attrs

* upgrade

* skip configflow for HACS

* message

* fixes

* fix list after restore

* remove unused const

* Add integrationhelper to requirements

* remove lines

* Fix some restore issues

* Fix migration

* Fix some api calls

* fix typo (#378)

* Wrap progress bar and add margin (#379)

* Add location for settings

* Fix typo

* Highlight active tab (#383)

* Highlight active tab

* Store active when in repository view

* Refactor logic and remove repository tab highlighting

* Fix issue with show_beta

* Fix version restore of HACS

* Version bump

* Force beta for hacs

* Adds AIOGitHubRatelimit exception

* Change exceptions

* flow: Unload HACS when there are issues

* Check for missing manifest

* Change exception

* Fix uninstall bug

* Change repository menu click trigger

* bump version number

* Don't warn about HACS

* Add info about the repositories as attribute

* Goto overview after uninstall

* Fix version select

* fix merge issue

* version bump

* Version change

+++ custom_components/hacs/handler/download.py
@@ -1,16 +1,14 @@
 """Download."""
+import os
 import gzip
-import logging
 import shutil
 
 import aiofiles
 import async_timeout
-
+from integrationhelper import Logger
 import backoff
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from ..hacsbase.exceptions import HacsNotSoBasicException
-
-_LOGGER = logging.getLogger("custom_components.hacs.download")
 
 
 @backoff.on_exception(backoff.expo, Exception, max_tries=5)
@@ -18,6 +16,7 @@ async def async_download_file(hass, url):
     """
     Download files, and return the content.
     """
+    logger = Logger("hacs.download.downloader")
     if url is None:
         return
 
@@ -25,7 +24,7 @@ async def async_download_file(hass, url):
     if "tags/" in url:
         url = url.replace("tags/", "")
 
-    _LOGGER.debug("Donwloading %s", url)
+    logger.debug(f"Donwloading {url}")
 
     result = None
 
@@ -47,14 +46,8 @@ async def async_download_file(hass, url):
 
 async def async_save_file(location, content):
     """Save files."""
-    if "-bundle" in location:
-        location = location.replace("-bundle", "")
-    if "lovelace-" in location.split("/")[-1]:
-        search = location.split("/")[-1]
-        replace = search.replace("lovelace-", "")
-        location = location.replace(search, replace)
-
-    _LOGGER.debug("Saving %s", location)
+    logger = Logger("hacs.download.save")
+    logger.debug(f"Saving {location}")
     mode = "w"
     encoding = "utf-8"
     errors = "ignore"
@@ -71,12 +64,15 @@ async def async_save_file(location, content):
             await outfile.write(content)
             outfile.close()
 
+        # Create gz for .js files
+        if os.path.isfile(location):
+            if location.endswith(".js") or location.endswith(".css"):
+                with open(location, "rb") as f_in:
+                    with gzip.open(location + ".gz", "wb") as f_out:
+                        shutil.copyfileobj(f_in, f_out)
+
     except Exception as error:  # pylint: disable=broad-except
         msg = "Could not write data to {} - {}".format(location, error)
-        _LOGGER.debug(msg)
+        logger.debug(msg)
 
-    # Create gz for .js files
-    if location.endswith(".js") or location.endswith(".css"):
-        with open(location, "rb") as f_in:
-            with gzip.open(location + ".gz", "wb") as f_out:
-                shutil.copyfileobj(f_in, f_out)
+    return os.path.exists(location)
